Still won't work

+++ poc/connection.cpp
@@ -1,13 +1,13 @@
 #include "connection.h"
 
-Connection::Connection(Neuron * startPoint, float level, float weight)
+Connection::Connection(Neuron * startPoint, int level, float weight)
 {
 	_startPoint = startPoint;
 	_level = level;
 	_weight = weight;
 }
 
-int Connection::getOutput(int input)
+float Connection::getOutput(float input)
 {
 	return input * _weight;
 }
